+++ frontend/kelp-me/tests/unit/organization/OrganizationComponent.spec.js
@@ -1,5 +1,5 @@
 import { mount } from '@vue/test-utils';
-import OrganizationComponent from './../../src/components/organization/OrganizationComponent.vue';
+import OrganizationComponent from '../../../src/components/organization/OrganizationComponent.vue';
 
 describe('Test para NoticiaComponent', () => {
     it('Test 1: Pintando props', () => {
